+++ Controllers/ChatController.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+using System.Web.Mvc;
+
+namespace ChhayaNirh.Controllers
+{
+    public class ChatController : Controller
+    {
+        // GET: Chat
+        public ActionResult Chat()
+        {
+            return View();
+        }
+    }
+}
