+++ Controllers/PostController.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+using System.Web.Mvc;
+
+namespace ChhayaNirh.Controllers
+{
+    public class PostController : Controller
+    {
+        // GET: Post
+        public ActionResult Post()
+        {
+            return View();
+        }
+    }
+}
