+++ Views/Profile/Profile.cshtml
@@ -28,8 +28,8 @@
 
                 <!-- Buttons -->
                 <div class="d-flex justify-content-center mt-4 gap-5">
-                    <a href="@Url.Action("EditProfile", "Home")" class="btn btn-primary">Edit Profile</a>
-                    <a href="@Url.Action("Logout", "Home")" class="btn btn-danger">Logout</a>
+                    <a href="@Url.Action("EditProfile", "Profile")" class="btn btn-primary">Edit Profile</a>
+                    <a href="@Url.Action("Logout", "Login")" class="btn btn-danger">Logout</a>
                 </div>
             </div>
         </div>
